Feat: Introduce `calculXp` function for accurate XP calculation
